+++ test/integration/connection/test-change-user.test.cjs
@@ -16,10 +16,10 @@ const onlyUsername = function (name) {
 };
 
 connection.query(
-  "CREATE USER IF NOT EXISTS 'changeuser1'@'%' IDENTIFIED BY 'changeuser1pass'",
+  "CREATE USER IF NOT EXISTS 'changeuser1'@'%' IDENTIFIED BY 'changeuser1pass'"
 );
 connection.query(
-  "CREATE USER IF NOT EXISTS 'changeuser2'@'%' IDENTIFIED BY 'changeuser2pass'",
+  "CREATE USER IF NOT EXISTS 'changeuser2'@'%' IDENTIFIED BY 'changeuser2pass'"
 );
 connection.query("GRANT ALL ON *.* TO 'changeuser1'@'%'");
 connection.query("GRANT ALL ON *.* TO 'changeuser2'@'%'");
@@ -48,7 +48,7 @@ connection.changeUser(
             assert.ifError(err);
             assert.deepEqual(
               onlyUsername(rows[0]['current_user()']),
-              'changeuser2',
+              'changeuser2'
             );
 
             connection.changeUser(
@@ -57,7 +57,7 @@ connection.changeUser(
                 password: 'changeuser1pass',
                 passwordSha1: Buffer.from(
                   'f961d39c82138dcec42b8d0dcb3e40a14fb7e8cd',
-                  'hex',
+                  'hex'
                 ), // sha1(changeuser1pass)
               },
               () => {
@@ -65,15 +65,15 @@ connection.changeUser(
                   assert.ifError(err);
                   assert.deepEqual(
                     onlyUsername(rows[0]['current_user()']),
-                    'changeuser1',
+                    'changeuser1'
                   );
                   connection.end();
                 });
-              },
+              }
             );
           });
-        },
+        }
       );
     });
-  },
+  }
 );