+++ test/integration/connection/test-execute-bind-date.test.cjs
@@ -17,7 +17,7 @@ connection.execute(
     }
     rows = _rows;
     connection.end();
-  },
+  }
 );
 
 process.on('exit', () => {