+++ test/integration/connection/test-invalid-date-result.test.cjs
@@ -44,7 +44,7 @@ connection.query(
         connection.end();
       });
     });
-  },
+  }
 );
 
 function isInvalidTime(t) {