+++ test/unit/connection/test-connection_config.test.cjs
@@ -11,7 +11,7 @@ assert.throws(
       ssl: true,
     }),
   (err) => err instanceof TypeError && err.message === expectedMessage,
-  'Error, the constructor accepts a boolean without throwing the right exception',
+  'Error, the constructor accepts a boolean without throwing the right exception'
 );
 
 assert.doesNotThrow(
@@ -19,7 +19,7 @@ assert.doesNotThrow(
     new ConnectionConfig({
       ssl: {},
     }),
-  'Error, the constructor accepts an object but throws an exception',
+  'Error, the constructor accepts an object but throws an exception'
 );
 
 assert.doesNotThrow(() => {
@@ -44,35 +44,35 @@ assert.doesNotThrow(() => {
 
 assert.strictEqual(
   ConnectionConfig.parseUrl(
-    String.raw`fml://test:pass!%40%24%25%5E%26*()word%3A@www.example.com/database`,
+    String.raw`fml://test:pass!%40%24%25%5E%26*()word%3A@www.example.com/database`
   ).password,
-  'pass!@$%^&*()word:',
+  'pass!@$%^&*()word:'
 );
 
 assert.strictEqual(
   ConnectionConfig.parseUrl(
-    String.raw`fml://user%40test.com:pass!%40%24%25%5E%26*()word%3A@www.example.com/database`,
+    String.raw`fml://user%40test.com:pass!%40%24%25%5E%26*()word%3A@www.example.com/database`
   ).user,
-  'user@test.com',
+  'user@test.com'
 );
 
 assert.strictEqual(
   ConnectionConfig.parseUrl(
-    String.raw`fml://test:pass@wordA@fe80%3A3438%3A7667%3A5c77%3Ace27%2518/database`,
+    String.raw`fml://test:pass@wordA@fe80%3A3438%3A7667%3A5c77%3Ace27%2518/database`
   ).host,
-  'fe80:3438:7667:5c77:ce27%18',
+  'fe80:3438:7667:5c77:ce27%18'
 );
 
 assert.strictEqual(
   ConnectionConfig.parseUrl(
-    String.raw`fml://test:pass@wordA@www.example.com/database`,
+    String.raw`fml://test:pass@wordA@www.example.com/database`
   ).host,
-  'www.example.com',
+  'www.example.com'
 );
 
 assert.strictEqual(
   ConnectionConfig.parseUrl(
-    String.raw`fml://test:pass@wordA@www.example.com/database%24`,
+    String.raw`fml://test:pass@wordA@www.example.com/database%24`
   ).database,
-  'database$',
+  'database$'
 );